+++ lib/src/base/text_selection.dart
@@ -1,16 +1,16 @@
+import 'dart:math' as math;
+
 import 'package:arna/arna.dart';
 import 'package:flutter/material.dart' show MaterialLocalizations;
-import 'package:flutter/rendering.dart';
+import 'package:flutter/rendering.dart' show TextSelectionPoint;
 
-// TODO: Convert clamp to clampDouble when it is landed. https://github.com/flutter/flutter/commit/64a0c19652a63292f7587a52f5f24656291ffd35
-
-/// Text selection controls.
-class _ArnaTextSelectionControls extends TextSelectionControls {
-  /// Desktop has no text selection handles.
+/// text selection controls.
+class ArnaTextSelectionControls extends TextSelectionControls {
+  /// Returns the size of the handle.
   @override
-  Size getHandleSize(double textLineHeight) => Size.zero;
+  Size getHandleSize(double textLineHeight) => const Size(Styles.handleSize, Styles.handleSize);
 
-  /// Builder for the Arna-style copy/paste text selection toolbar.
+  /// Builder for Arna-style copy/paste text selection toolbar.
   @override
   Widget buildToolbar(
     BuildContext context,
@@ -23,28 +23,61 @@ class _ArnaTextSelectionControls extends TextSelectionControls {
     Offset? lastSecondaryTapDownPosition,
   ) {
     return _ArnaTextSelectionControlsToolbar(
-      clipboardStatus: clipboardStatus,
-      endpoints: endpoints,
       globalEditableRegion: globalEditableRegion,
+      textLineHeight: textLineHeight,
+      selectionMidpoint: selectionMidpoint,
+      endpoints: endpoints,
+      delegate: delegate,
+      clipboardStatus: clipboardStatus,
       handleCut: canCut(delegate) ? () => handleCut(delegate) : null,
       handleCopy: canCopy(delegate) ? () => handleCopy(delegate) : null,
       handlePaste: canPaste(delegate) ? () => handlePaste(delegate) : null,
       handleSelectAll: canSelectAll(delegate) ? () => handleSelectAll(delegate) : null,
-      selectionMidpoint: selectionMidpoint,
-      lastSecondaryTapDownPosition: lastSecondaryTapDownPosition,
-      textLineHeight: textLineHeight,
     );
   }
 
-  /// Builds the text selection handles, but desktop has none.
+  /// Builder for Arna-style text selection handles.
   @override
-  Widget buildHandle(BuildContext context, TextSelectionHandleType type, double textLineHeight, [VoidCallback? onTap]) {
-    return const SizedBox.shrink();
+  Widget buildHandle(BuildContext context, TextSelectionHandleType type, double textHeight, [VoidCallback? onTap]) {
+    final Color handleColor = ArnaDynamicColor.resolve(ArnaColors.iconColor, context);
+    final Widget handle = SizedBox(
+      width: Styles.handleSize,
+      height: Styles.handleSize,
+      child: CustomPaint(
+        painter: _ArnaTextSelectionHandlePainter(color: handleColor),
+        child: GestureDetector(
+          onTap: onTap,
+          behavior: HitTestBehavior.translucent,
+        ),
+      ),
+    );
+
+    // [handle] is a circle, with a rectangle in the top left quadrant of that circle (an onion pointing to 10:30). We
+    // rotate [handle] to point straight up or up-right depending on the handle type.
+    switch (type) {
+      case TextSelectionHandleType.left: // points up-right
+        return Transform.rotate(angle: math.pi / 2.0, child: handle);
+      case TextSelectionHandleType.right: // points up-left
+        return handle;
+      case TextSelectionHandleType.collapsed: // points up
+        return Transform.rotate(angle: math.pi / 4.0, child: handle);
+    }
   }
 
-  /// Gets the position for the text selection handles, but desktop has none.
+  /// Gets anchor for Arna-style text selection handles.
+  ///
+  /// See [TextSelectionControls.getHandleAnchor].
   @override
-  Offset getHandleAnchor(TextSelectionHandleType type, double textLineHeight) => Offset.zero;
+  Offset getHandleAnchor(TextSelectionHandleType type, double textLineHeight) {
+    switch (type) {
+      case TextSelectionHandleType.left:
+        return const Offset(Styles.handleSize, 0);
+      case TextSelectionHandleType.right:
+        return Offset.zero;
+      case TextSelectionHandleType.collapsed:
+        return const Offset(Styles.handleSize / 2, -4);
+    }
+  }
 
   @override
   bool canSelectAll(TextSelectionDelegate delegate) {
@@ -54,41 +87,35 @@ class _ArnaTextSelectionControls extends TextSelectionControls {
         value.text.isNotEmpty &&
         !(value.selection.start == 0 && value.selection.end == value.text.length);
   }
-
-  @override
-  void handleSelectAll(TextSelectionDelegate delegate) {
-    super.handleSelectAll(delegate);
-    delegate.hideToolbar();
-  }
 }
 
 /// Text selection controls.
-final TextSelectionControls arnaTextSelectionControls = _ArnaTextSelectionControls();
+final TextSelectionControls arnaTextSelectionControls = ArnaTextSelectionControls();
 
-// Generates the child that's passed into ArnaTextSelectionToolbar.
+// The highest level toolbar widget, built directly by buildToolbar.
 class _ArnaTextSelectionControlsToolbar extends StatefulWidget {
   /// Creates the child that's passed into ArnaTextSelectionToolbar.
   const _ArnaTextSelectionControlsToolbar({
     required this.clipboardStatus,
+    required this.delegate,
     required this.endpoints,
     required this.globalEditableRegion,
-    required this.handleCopy,
     required this.handleCut,
+    required this.handleCopy,
     required this.handlePaste,
     required this.handleSelectAll,
     required this.selectionMidpoint,
     required this.textLineHeight,
-    required this.lastSecondaryTapDownPosition,
   });
 
   final ClipboardStatusNotifier? clipboardStatus;
+  final TextSelectionDelegate delegate;
   final List<TextSelectionPoint> endpoints;
   final Rect globalEditableRegion;
-  final VoidCallback? handleCopy;
   final VoidCallback? handleCut;
+  final VoidCallback? handleCopy;
   final VoidCallback? handlePaste;
   final VoidCallback? handleSelectAll;
-  final Offset? lastSecondaryTapDownPosition;
   final Offset selectionMidpoint;
   final double textLineHeight;
 
@@ -97,7 +124,8 @@ class _ArnaTextSelectionControlsToolbar extends StatefulWidget {
 }
 
 /// The [State] for an [_ArnaTextSelectionControlsToolbar].
-class _ArnaTextSelectionControlsToolbarState extends State<_ArnaTextSelectionControlsToolbar> {
+class _ArnaTextSelectionControlsToolbarState extends State<_ArnaTextSelectionControlsToolbar>
+    with TickerProviderStateMixin {
   // Inform the widget that the value of clipboardStatus has changed.
   void _onChangedClipboardStatus() => setState(() {});
 
@@ -110,9 +138,9 @@ class _ArnaTextSelectionControlsToolbarState extends State<_ArnaTextSelectionCon
   @override
   void didUpdateWidget(_ArnaTextSelectionControlsToolbar oldWidget) {
     super.didUpdateWidget(oldWidget);
-    if (oldWidget.clipboardStatus != widget.clipboardStatus) {
-      oldWidget.clipboardStatus?.removeListener(_onChangedClipboardStatus);
+    if (widget.clipboardStatus != oldWidget.clipboardStatus) {
       widget.clipboardStatus?.addListener(_onChangedClipboardStatus);
+      oldWidget.clipboardStatus?.removeListener(_onChangedClipboardStatus);
     }
   }
 
@@ -124,23 +152,38 @@ class _ArnaTextSelectionControlsToolbarState extends State<_ArnaTextSelectionCon
 
   @override
   Widget build(BuildContext context) {
-    // Don't render the menu until the state of the clipboard is known.
+    // If there are no buttons to be shown, don't render anything.
+    if (widget.handleCut == null &&
+        widget.handleCopy == null &&
+        widget.handlePaste == null &&
+        widget.handleSelectAll == null) {
+      return const SizedBox.shrink();
+    }
+    // If the paste button is desired, don't render anything until the state of the clipboard is known, since it's used
+    // to determine if paste is shown.
     if (widget.handlePaste != null && widget.clipboardStatus?.value == ClipboardStatus.unknown) {
-      return const SizedBox(width: 0.0, height: 0.0);
+      return const SizedBox.shrink();
     }
 
-    assert(debugCheckHasMediaQuery(context));
-    final MediaQueryData mediaQuery = MediaQuery.of(context);
+    // Calculate the positioning of the menu. It is placed above the selection if there is enough room, or otherwise
+    // below.
+    final TextSelectionPoint startTextSelectionPoint = widget.endpoints[0];
+    final TextSelectionPoint endTextSelectionPoint =
+        widget.endpoints.length > 1 ? widget.endpoints[1] : widget.endpoints[0];
+    final double topAmountInEditableRegion = startTextSelectionPoint.point.dy - widget.textLineHeight;
+    final double anchorTop = math.max(topAmountInEditableRegion, 0) + widget.globalEditableRegion.top - Styles.padding;
 
-    final Offset midpointAnchor = Offset(
-      (widget.selectionMidpoint.dx - widget.globalEditableRegion.left).clamp(
-        mediaQuery.padding.left,
-        mediaQuery.size.width - mediaQuery.padding.right,
-      ),
-      widget.selectionMidpoint.dy - widget.globalEditableRegion.top,
+    final Offset anchorAbove = Offset(
+      widget.globalEditableRegion.left + widget.selectionMidpoint.dx,
+      anchorTop,
+    );
+    final Offset anchorBelow = Offset(
+      widget.globalEditableRegion.left + widget.selectionMidpoint.dx,
+      widget.globalEditableRegion.top + endTextSelectionPoint.point.dy + Styles.handleSize - 2,
     );
 
     final MaterialLocalizations localizations = MaterialLocalizations.of(context);
+
     final List<Widget> items = <Widget>[];
 
     void addToolbarButton(IconData icon, String text, VoidCallback onPressed) {
@@ -188,9 +231,22 @@ class _ArnaTextSelectionControlsToolbarState extends State<_ArnaTextSelectionCon
     }
 
     return _ArnaTextSelectionToolbar(
-      anchor: widget.lastSecondaryTapDownPosition ?? midpointAnchor,
+      anchorAbove: anchorAbove,
+      anchorBelow: anchorBelow,
       children: items,
     );
+
+    // return ArnaTextSelectionToolbar(
+    //   anchorAbove: anchorAbove,
+    //   anchorBelow: anchorBelow,
+    //   children: itemDatas.asMap().entries.map((MapEntry<int, _TextSelectionToolbarItemData> entry) {
+    //     return TextSelectionToolbarTextButton(
+    //       padding: TextSelectionToolbarTextButton.getPadding(entry.key, itemDatas.length),
+    //       onPressed: entry.value.onPressed,
+    //       child: Text(entry.value.label),
+    //     );
+    //   }).toList(),
+    // );
   }
 }
 
@@ -199,19 +255,22 @@ class _ArnaTextSelectionControlsToolbarState extends State<_ArnaTextSelectionCon
 /// Typically displays buttons for text manipulation, e.g. copying and pasting text.
 ///
 /// Tries to position itself as closely as possible to [anchor] while remaining fully on-screen.
-///
-/// See also:
-///
-///  * [_ArnaTextSelectionControls.buildToolbar], where this is used by default to build an Arna-style toolbar.
 class _ArnaTextSelectionToolbar extends StatelessWidget {
   /// Creates an instance of _ArnaTextSelectionToolbar.
   const _ArnaTextSelectionToolbar({
-    required this.anchor,
+    required this.anchorAbove,
+    required this.anchorBelow,
     required this.children,
   }) : assert(children.length > 0);
 
-  /// The point at which the toolbar will attempt to position itself as closely as possible.
-  final Offset anchor;
+  /// The focal point above which the toolbar attempts to position itself.
+  ///
+  /// If there is not enough room above before reaching the top of the screen, then the toolbar will position itself
+  /// below [anchorBelow].
+  final Offset anchorAbove;
+
+  /// The focal point below which the toolbar attempts to position itself, if it doesn't fit above [anchorAbove].
+  final Offset anchorBelow;
 
   /// {@macro flutter.material.TextSelectionToolbar.children}
   ///
@@ -225,6 +284,8 @@ class _ArnaTextSelectionToolbar extends StatelessWidget {
     final MediaQueryData mediaQuery = MediaQuery.of(context);
 
     final double paddingAbove = mediaQuery.padding.top + Styles.padding;
+    final double availableHeight = anchorAbove.dy - paddingAbove;
+    final bool fitsAbove = Styles.buttonSize <= availableHeight;
     final Offset localAdjustment = Offset(Styles.padding, paddingAbove);
 
     return Padding(
@@ -235,8 +296,10 @@ class _ArnaTextSelectionToolbar extends StatelessWidget {
         Styles.padding,
       ),
       child: CustomSingleChildLayout(
-        delegate: DesktopTextSelectionToolbarLayoutDelegate(
-          anchor: anchor - localAdjustment,
+        delegate: TextSelectionToolbarLayoutDelegate(
+          anchorAbove: anchorAbove - localAdjustment,
+          anchorBelow: anchorBelow - localAdjustment,
+          fitsAbove: fitsAbove,
         ),
         child: ArnaCard(
           child: Wrap(children: children),
@@ -272,4 +335,27 @@ class _ArnaTextSelectionToolbarButton extends StatelessWidget {
       tooltipMessage: label,
     );
   }
+}
+
+/// Draws a single text selection handle which points up and to the left.
+class _ArnaTextSelectionHandlePainter extends CustomPainter {
+  /// Draws a single text selection handle.
+  _ArnaTextSelectionHandlePainter({required this.color});
+
+  final Color color;
+
+  @override
+  void paint(Canvas canvas, Size size) {
+    final Paint paint = Paint()..color = color;
+    final double radius = size.width / 2.0;
+    final Rect circle = Rect.fromCircle(center: Offset(radius, radius), radius: radius);
+    final Rect point = Rect.fromLTWH(0.0, 0.0, radius, radius);
+    final Path path = Path()
+      ..addOval(circle)
+      ..addRect(point);
+    canvas.drawPath(path, paint);
+  }
+
+  @override
+  bool shouldRepaint(_ArnaTextSelectionHandlePainter oldPainter) => color != oldPainter.color;
 }